Removed redundant copies and optimized data transfers
