+++ omp/README
