+++ omp/lulesh-comm.cc
