+++ omp/lulesh-viz.cc
